refactor(frontend): renamed messageComp to channel
